+++ AI5R-STUDIO/dashboard/src/design-system/inspector/index.js
@@ -1,0 +1,4 @@
+export { default as Inspector } from "./Inspector";
+export { default as InspectorHeader } from "./InspectorHeader";
+export { default as InspectorSection } from "./InspectorSection";
+export { default as InspectorField } from "./InspectorField";